[O2B-693] Refactor table exports
