+++ lib/public/components/common/table/headers.js
@@ -21,7 +21,7 @@ import { iconCaretTop, iconCaretBottom, iconX } from '/js/src/icons.js';
  * @param {object|null} filterModel The model storing the filtering state
  * @return {vnode} An array of rows containing all given header values with specific cell sizes
  */
-const headers = (columnsDescriptions, sortModel, filterModel) => {
+export const headers = (columnsDescriptions, sortModel, filterModel) => {
     const entries = Object.entries(columnsDescriptions).filter(([_, columnDescription]) => columnDescription.visible);
     const isInlineFilter = entries.some(([_, value]) => value.inlineFilter);
 
@@ -88,5 +88,3 @@ const headers = (columnsDescriptions, sortModel, filterModel) => {
         return h(`th#${key}.${classes.filter((item) => Boolean(item)).join(' ')}`, attributes, content);
     })));
 };
-
-export default headers;
