+++ lib/public/components/common/table/rows.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import { wrapWithBalloon } from '../../utilities/wrapWithBalloon.js';
+import { wrapWithBalloon } from '../../../utilities/wrapWithBalloon.js';
 
 /**
  * Renders a single cell with content
@@ -23,7 +23,7 @@ import { wrapWithBalloon } from '../../utilities/wrapWithBalloon.js';
  * @param {Object} entry The specific object
  * @return {Vnode} A single table cell containing (formatted) text
  */
-const content = (key, value, text, rowId, entry) => {
+const cell = (key, value, text, rowId, entry) => {
     value = {
         format: (text) => text,
         balloon: false,
@@ -57,7 +57,7 @@ const content = (key, value, text, rowId, entry) => {
  * @param {string} row string
  * @return {vnode} A filled array of rows based on the given data and keys
  */
-const rows = (data, columnsDescriptions, rowConfigurationCallback, row = '') => {
+export const rows = (data, columnsDescriptions, rowConfigurationCallback, row = '') => {
     const idKey = Object.keys(columnsDescriptions).find((key) => columnsDescriptions[key] && columnsDescriptions[key].primary);
 
     return h('tbody', data.map((entry) => {
@@ -67,11 +67,9 @@ const rows = (data, columnsDescriptions, rowConfigurationCallback, row = '') => 
             ...rowConfigurationCallback && rowConfigurationCallback(entry),
         }, Object.entries(columnsDescriptions).reduce((accumulator, [key, value]) => {
             if (value.visible) {
-                accumulator.push(content(key, value, entry[key], rowId, entry));
+                accumulator.push(cell(key, value, entry[key], rowId, entry));
             }
             return accumulator;
         }, []));
     }));
 };
-
-export default rows;
