+++ lib/public/components/common/table/table.js
@@ -12,10 +12,10 @@
  */
 
 import { h, RemoteData } from '/js/src/index.js';
-import headers from './headers.js';
-import rows from './rows.js';
-import spinner from '../common/spinner.js';
-import errorAlert from '../common/errorAlert.js';
+import { headers } from './headers.js';
+import { rows } from './rows.js';
+import spinner from '../spinner.js';
+import errorAlert from '../errorAlert.js';
 
 /**
  * @callback rowConfigurationCallback
@@ -36,13 +36,20 @@ import errorAlert from '../common/errorAlert.js';
  * @param {string} rowClasses the css classes to apply to each row (separated by a dot)
  * @returns {vnode} Return the total view of the table to rendered
  */
-const table = (data, columnsDescriptions, { sortModel = null, filterModel = null } = {}, rowConfigurationCallback = null, rowClasses = '') => {
+export const table = (
+    data,
+    columnsDescriptions,
+    { sortModel = null, filterModel = null } = {},
+    rowConfigurationCallback = null,
+    rowClasses = '',
+) => {
     let content = data;
     if (Array.isArray(data)) {
         content = RemoteData.success(data);
     }
 
-    const columnsCount = Object.values(columnsDescriptions).filter((columnDescription) => columnDescription.visible).length;
+    const columnsCount = Object.values(columnsDescriptions)
+        .filter((columnDescription) => columnDescription.visible).length;
 
     return h(
         'table.table.table-hover.shadow-level1.no-z-index',
@@ -67,5 +74,3 @@ const table = (data, columnsDescriptions, { sortModel = null, filterModel = null
         ],
     );
 };
-
-export default table;
