+++ lib/public/views/Flps/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import { amountSelector, pageSelector } from '../../../components/Pagination/index.js';
 import { flpsActiveColumns } from '../ActiveColumns/flpsActiveColumns.js';
