+++ lib/public/views/Home/Overview/index.js
@@ -14,7 +14,7 @@
 import { h } from '/js/src/index.js';
 import { homeOverviewLogsActiveColumns } from '../../Logs/ActiveColumns/logsActiveColumns.js';
 import { homeOverviewRunsActiveColumns } from '../../Runs/ActiveColumns/runsActiveColumns.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 
 const TABLEROW_HEIGHT = 46;
 // Estimate of the navbar and pagination elements height total; Needs to be updated in case of changes;
