+++ lib/public/views/LhcFills/Detail/lhcFillDetailsComponent.js
@@ -13,7 +13,7 @@
 
 import { h } from '/js/src/index.js';
 import spinner from '../../../components/common/spinner.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import { runsActiveColumns } from '../../Runs/ActiveColumns/runsActiveColumns.js';
 import { lhcFillDisplayConfiguration } from './lhcFillDisplayConfiguration.js';
