+++ lib/public/views/LhcFills/Overview/index.js
@@ -13,7 +13,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import { amountSelector, pageSelector } from '../../../components/Pagination/index.js';
 import { lhcFillsActiveColumns } from '../ActiveColumns/lhcFillsActiveColumns.js';
 import { estimateDisplayableRowsCount } from '../../../utilities/estimateDisplayableRowsCount.js';
