+++ lib/public/views/Logs/Logs.js
@@ -16,7 +16,7 @@ import { setMarkDownBox } from '../../components/common/markdown.js';
 import { removeElement } from '../../utilities/removeElement.js';
 import { TagPickerModel } from '../../components/pickers/tag/TagPickerModel.js';
 import { TagFilterModel } from '../../components/Filters/common/TagFilterModel.js';
-import { SortModel } from '../../components/Table/SortModel.js';
+import { SortModel } from '../../components/common/table/SortModel.js';
 import { debounce, INPUT_DEBOUNCE_TIME } from '../../utilities/debounce.js';
 import { taggedEventRegistry } from '../../utilities/taggedEventRegistry.js';
 import { FILTER_PANEL_CLICK_TAG } from '../../components/Filters/filtersConstants.js';
