+++ lib/public/views/Logs/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import { amountSelector, pageSelector } from '../../../components/Pagination/index.js';
 import { logsActiveColumns } from '../ActiveColumns/logsActiveColumns.js';
 import { filtersPanel } from '../../../components/Filters/common/filtersPanel.js';
