+++ lib/public/views/Runs/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import { amountSelector, pageSelector } from '../../../components/Pagination/index.js';
 import { runsActiveColumns } from '../ActiveColumns/runsActiveColumns.js';
 import { exportRunsModal } from './exportRunsModal.js';
