+++ lib/public/views/Subsystems/Details/index.js
@@ -13,7 +13,7 @@
 
 import { h } from '/js/src/index.js';
 import spinner from '../../../components/common/spinner.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import targetURL from '../../../utilities/targetURL.js';
 import { logsActiveColumns } from '../../Logs/ActiveColumns/logsActiveColumns.js';
