+++ lib/public/views/Subsystems/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 
 const ACTIVE_COLUMNS = {
     id: {
