+++ lib/public/views/Tags/Details/index.js
@@ -14,7 +14,7 @@
 import { h } from '/js/src/index.js';
 import errorAlert from '../../../components/common/errorAlert.js';
 import spinner from '../../../components/common/spinner.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import targetURL from '../../../utilities/targetURL.js';
 import { reducedLogsActiveColumns } from '../../Logs/ActiveColumns/reducedLogsActiveColumns.js';
 import tagDetail from '../../../components/TagDetail/index.js';
