+++ lib/public/views/Tags/Overview/index.js
@@ -12,7 +12,7 @@
  */
 
 import { h } from '/js/src/index.js';
-import table from '../../../components/Table/index.js';
+import { table } from '../../../components/common/table/table.js';
 import { tagsActiveColumns } from '../ActiveColumns/tagsActiveColumns.js';
 
 /**
